Separate `IHttpRequest` base class without curl dependency
Create a separate `IHttpRequest` base class for HTTP requests in the `src/engine/http.cpp/h` files independent from the curl-based implementation.

Move all details specific to the curl-based implementation to the `CHttpCurl` and `CHttpRequestCurl` classes in the `src/engine/shared/http_curl.cpp/h` files. No other files should depend on these implementation details anymore, which allows adding `#include <curl/curl.h>` in `http_curl.h` and using the curl-specific types. Replace all existing usages of the `CHttpRequest` class with the base class `IHttpRequest`.

When building for Emscripten, this will allow replacing the non-functional curl-based implementation with the Emscripten Fetch API.

It is not intended anymore for either the `IHttpRequest` or the `CHttpRequestCurl` class to be subclassed by users of the HTTP request interface. Use composition instead of inheritance for old subclasses `CCommunityIconDownloadJob` and `CUpdaterFetchTask` of `CHttpRequest`. Instead of having to override the HTTP request class for the `OnProgress` and `OnCompletion` callbacks, move these to a separate interface `IHttpRequest::IProgressCallback` that can be set when creating an HTTP request.

Make the global function `HasIpresolveBug` a virtual member function of `IHttp`.

Rename the config variable `dbg_curl` to `dbg_http`.

# Conflicts:
#	src/engine/shared/http.cpp
#	src/engine/shared/http.h
