+++ src/engine/client/updater.cpp
@@ -7,7 +7,7 @@
 #include <engine/client.h>
 #include <engine/engine.h>
 #include <engine/external/json-parser/json.h>
-#include <engine/shared/http.h>
+#include <engine/http.h>
 #include <engine/shared/json.h>
 #include <engine/storage.h>
 
@@ -20,18 +20,19 @@
 #include <sys/stat.h>
 #endif
 
-class CUpdaterFetchTask : public CHttpRequest
+class CUpdaterFetchTask : public IHttpRequest::IProgressCallback
 {
 	char m_aBuf[256];
 	CUpdater *m_pUpdater;
-
-	void OnProgress() override;
+	std::shared_ptr<IHttpRequest> m_pHttpRequest;
 
 protected:
+	void OnProgress() override;
 	void OnCompletion(EHttpState State) override;
 
 public:
 	CUpdaterFetchTask(CUpdater *pUpdater, const char *pFile, const char *pDestPath);
+	std::shared_ptr<IHttpRequest> HttpRequest() { return m_pHttpRequest; }
 };
 
 // addition of '/' to keep paths intact, because EscapeUrl() (using curl_easy_escape) doesn't do this
@@ -118,23 +119,24 @@ static bool SetExecutableBit(const char *pPath)
 #endif
 
 CUpdaterFetchTask::CUpdaterFetchTask(CUpdater *pUpdater, const char *pFile, const char *pDestPath) :
-	CHttpRequest(GetUpdaterUrl(m_aBuf, sizeof(m_aBuf), pFile)),
 	m_pUpdater(pUpdater)
 {
 	char aDestination[IO_MAX_PATH_LENGTH];
 	FormatUpdaterDestPath(aDestination, sizeof(aDestination), pFile, pDestPath);
-	WriteToFile(pUpdater->m_pStorage, aDestination, -2);
+	m_pHttpRequest = CreateHttpRequest(GetUpdaterUrl(m_aBuf, sizeof(m_aBuf), pFile));
+	m_pHttpRequest->WriteToFile(pUpdater->m_pStorage, aDestination, -2);
+	m_pHttpRequest->SetProgressCallback(this);
 }
 
 void CUpdaterFetchTask::OnProgress()
 {
 	const CLockScope LockScope(m_pUpdater->m_Lock);
-	m_pUpdater->m_Percent = Progress();
+	m_pUpdater->m_Percent = m_pHttpRequest->Progress();
 }
 
 void CUpdaterFetchTask::OnCompletion(EHttpState State)
 {
-	if(!str_comp(fs_filename(Dest()), "update.json"))
+	if(!str_comp(fs_filename(m_pHttpRequest->Dest()), "update.json"))
 	{
 		if(State == EHttpState::DONE)
 			m_pUpdater->SetCurrentState(IUpdater::GOT_MANIFEST);
@@ -195,8 +197,8 @@ void CUpdater::FetchFile(const char *pFile, const char *pDestPath)
 {
 	const CLockScope LockScope(m_Lock);
 	m_pCurrentTask = std::make_shared<CUpdaterFetchTask>(this, pFile, pDestPath);
-	str_copy(m_aStatus, m_pCurrentTask->Dest());
-	m_pHttp->Run(m_pCurrentTask);
+	str_copy(m_aStatus, m_pCurrentTask->HttpRequest()->Dest());
+	m_pHttp->Run(m_pCurrentTask->HttpRequest());
 }
 
 bool CUpdater::MoveFile(const char *pFile)
@@ -387,11 +389,12 @@ void CUpdater::RunningUpdate()
 {
 	if(m_pCurrentTask)
 	{
-		if(!m_pCurrentTask->Done())
+		if(!m_pCurrentTask->HttpRequest()->Done())
 		{
 			return;
 		}
-		else if(m_pCurrentTask->State() == EHttpState::ERROR || m_pCurrentTask->State() == EHttpState::ABORTED)
+		else if(m_pCurrentTask->HttpRequest()->State() == EHttpState::ERROR ||
+			m_pCurrentTask->HttpRequest()->State() == EHttpState::ABORTED)
 		{
 			SetCurrentState(IUpdater::FAIL);
 		}
