+++ src/game/client/components/censor.h
@@ -4,7 +4,6 @@
 
 #include <engine/console.h>
 #include <engine/shared/config.h>
-#include <engine/shared/http.h>
 #include <engine/shared/jobs.h>
 
 #include <game/client/component.h>
@@ -17,6 +16,8 @@
 void CensorReplaceWords(char *pBuffer, const std::vector<std::string> &vWords, char Replacement);
 
 /*
+class IHttpRequest;
+
 class CCensor : public CComponent
 {
 private:
@@ -39,7 +40,7 @@ private:
 		char m_aUrl[IO_MAX_PATH_LENGTH];
 		char m_aSaveFilePath[IO_MAX_PATH_LENGTH];
 		CLock m_Lock;
-		std::shared_ptr<CHttpRequest> m_pGetRequest;
+		std::shared_ptr<IHttpRequest> m_pGetRequest;
 	};
 
 	std::shared_ptr<CCensorListDownloadJob> m_pCensorListDownloadJob = nullptr;
